Flattened PostalError
